requirements.txt
Django
djangorestframework
future
pytz
